make unit_convert work in temp

+++ lib/src/constant/unit/capacitance_unit_details.dart
@@ -30,7 +30,7 @@ final _otherUnits = {
       ],
     ),
     CapacitanceUnit.abFarad,
-    conversionFactor: pow(10, 9),
+    conversionFactor: pow(10, 9).toDouble(),
     variation: true,
   ),
 };
